remove assetsDir backwards compatibility (#3341)
the `assetsDir` option for the `@Component` decorator was deprecated
some time ago in favor of the `assetsDirs` option, but we have retained
backwards-compatibility with the old option nonetheless. this commit
removes that backwards compatibility.

STENCIL-410: Remove Backwards Compatibility for assetDir Field on @Component

BREAKING CHANGE: setting `assetsDir` on a component will no longer work.
Users should migrate to `assetsDirs` instead.