fix(sys): make NodeLazyRequire complain if package versions aren't right
This updates a bit of logic in `NodeLazyRequire.ensure` to check that
the installed versions of packages are within the specified version
range, i.e. that `minVersion <= installedVersion <= maxVersion`.

This commit also adds tests for that module.

STENCIL-391: bug: @stencil/core does not throw error when missing
jest/jest-cli deps in a rush/pnpm monorepo