test(bundler): add vite test
this commit adds an application that is built using vite to the bundler
test directory. it consumes a small stencil library build using the
`dist` output target, and verifies that the application can load the web
component when the application has been built using vite.

this commit does not add the infrastructure to run the tests themselves.
that will be completed in a follow on PR