test(bundler): add parcel test
this commit adds an application that is built using parcel (v2) to the
bundler test directory. this application consumes a small stencil
library built using the `dist` output target. the included test verifies
the applicaiton can load the web component when the application has been
built using parcel.

this commit also includes the infrastructure updates necessary to run
the test using karma.

in order to get the test to work, as shell script was added to
reorganize the structure of the output directory. regardless of the
structure of its input, parcel will always flatten its output to a
single directory. as a result, the src attributes on script tags will
always start with a leading forward slash. this is problematic when
attempting to run these tests in karma, as karma will try to resolve
these files from the root of its server. to circumvent this,
`restructure.sh` moves all files (other than the index.html entrypoint)
into a sub-directory, `parcel-assets`, and rewrites `index.html`'s src
attributes accordingly. this allows us to serve the files in a way that:
in order to get the test to work, as shell script was added to
reorganize the structure of the output directory. regardless of the
structure of its input, parcel will always flatten its output to a
single directory. as a result, the src attributes on script tags will
always start with a leading forward slash. this is problematic when
attempting to run these tests in karma, as karma will try to resolve
these files from the root of its server. to circumvent this,
`restructure.sh` moves all files (other than the index.html entrypoint)
into a sub-directory, `parcel-assets`, and rewrites `index.html`'s src
attributes accordingly. this allows us to serve the files in a way that:
1. doesn't collide with other bundler tests
2. allows files to be correctly resolved for parcel

this is admittedly a bit of a hack. while i don't think this completely
undermines the test (since we're rewriting the output of parcel), it
does compromise the spirit a bit. if we decide _not_ to land this, i'm
happy to drop this commit and just continue on with the existing
(vite-based) one