feat(compiler): export all built components from index.js w/ dist-custom-elements
This makes a change to the `dist-custom-elements` output target code
which adds exports from the generated `index.js` file for all of the
components included in the build. This should help to enable Stencil
users to migrate away from the `dist-custom-elements-bundle` output
target, which we're planning to remove in the future (see #3136 for
details and discussion on that).

In order to enable this we don't need to make a particularly large
change. The index chunk which we generate and pass to Rollup is just
amended to include some references to the other modules we declare (one
per each component), and Rollup takes care of resolving that into
actual, bundled concrete modules.

As part of making this change there is also a new test file added to
exercise the functions for the `dist-custom-elements` output target,
which necessitated improving our mocks a little bit. This should help us
to test the rest of the output target code in the future.

STENCIL-332 Investigate re-exporting components via index.js