refactor(docs): remove strictNullChecks errors for vscode
this commit updates the vscode documentation generation file to remove
existing `strictNullChecks` violations. the goal for these changes were
to improve type correctness, without changing existing functionality

`getReferences()` - the changes to this function accomodate that the
fields that are received from the output target is not required (e.g.
`sourceCodeBaseUrl` is optional) or to work with fields that could be
`undefined`

`serializeAttribute()` - two custom types are added here to validate
that the 'attr' field exists on the provided prop documentaiton (and
should, based on a `filter` that was performed prior to invocation)