feat(compiler): generate component custom event types with HTML target (#3296)
with this commit, the compiler will generate new exported
interfaces in `components.d.ts` for all web components that
have custom events. the generated interfaces include the
typings to the generated HTML element type for the event target
and accepts a generic for the custom event detail. the compiler
will generate the web component custom event type with the
new type, allowing usage with the custom event and the generated
HTML element type for the target.