refactor(docs): document vscode output target  (#3359)
this commit adds jsdoc to the file responsible for generating custom
data for vscode (custom data allows vs code users to augment the html
element namespace with additional elements - e.g. custom elements).

it removes previously anonymous types with named ones
to aid in mapping the types back to the JSON schema/specification for
generating custom data. at this time, these types have not been tied
back to the types exported by the vs code html language service -
further investigation into the value of doing so is necessary and out of
the scope of this commit.

in addition to the removal of previously anonymous types, additional
return types, type guards, etc. have been added. these have been added
to improve some of the correctness of this file and to improve the
readability (by shortening the mental work needed to understand the
types that are being iterated over).

this commit updates the vscode documentation generation file to remove
existing `strictNullChecks` violations. the goal for these changes were
to improve type correctness, without changing existing functionality