chore(github): add strictNullChecks CI check (#3356)
This is basically a CI / tooling PR which adds a CI action that runs tsc with --strictNullchecks on both the branch for a given PR and on main and reports via a comment on the PR 1) differences in error count between the two and 2) some statistics about errors on the PR (which files have the most, etc).

The idea is that this will firstly discourage us from checking in code which adds null check errors, and then also encourage us to make changes which decrease the number of errors. We currently have 2188, which while kind of a lot is not so many, and a few dozen PRs which each clear up 50-100 or so will get us to where we want to be. So the hope is that by increasing visibility with this CI action we'll keep on improving things in that regard.

This check is intended to be removed in the future when we fix all the errors and are then in a position to turn strictNullChecks on all the time. We could also use it for enabling other compiler options which increase the error count as well.