chore(tests): fix end to end .d.ts file
this commit checks in changes to `test/end-to-end/src/components.d.ts`
that are generated starting with 846740fa1f074e401fa20be90d64e87f8db99c88
(https://github.com/ionic-team/stencil/pull/3296)

we missed this as a result of relying on ci to run the end to end tests
for us, rather than ever running them locally. the changes found in this
commit were generated by running `npm run test.end-to-end` on the commit
immediately preceding this one.

to avoid issues like this in the future, stencil-451 has been created
and added to the backlog