fix(compiler): prevent double full builds
this commit removes a file watcher from the process of performing a full
build of stencil.

root cause analysis:

note: this issue was easier to reproduce on an older machine with the
following specs:
- OS: Ubuntu 20.04.4 LTS, x86_64
- CPU: i5-4300U - 4cores, 2.9GHz
- RAM: 8GB

in some cases, it would reproduce on more modern hardware, such as an m1
macbook air, but was far less consistent.

the `src/compiler/types/generate-app-types.ts#generateAppTypes` function
is responsible for writing the `src/components.d.ts` file to disk. the
bug would manifest when the aforementioned function performed a write
of the type declaration file _and_ the write would occupy some period
of time that exceeded the time it would take for stencil to perform a
compilation.

`src/compiler/transpile/create-build-program.ts#createTsBuildProgram()`
is the function/file where this behavior would occur. the file watcher
that was removed in this commit would be created as a part of the
`ts.System` that would be used in the creation of a `ts.CompilerHost`
(and subsequently, a `ts.Program`). stencil overrides the
`afterProgramCreate` method on the `ts.WatchCompilerHost` interface,
which invokes a build callback - this callback is reponsible for
invoking `generateAppTypes()` via the following call chain:

`src/compiler/build/full-build.ts#createFullBuild#onBulid()` ->
`src/compiler/build/build.ts#build()` ->
`src/compiler/transpile/run-program.ts#runTsProgram()` ->
`src/compiler/types/generate-app-typese-.ts#generateAppTypes()`

the write to disk would trigger the now removed file watcher via an
async event, which would then kick off an additional compilation. this
second compilation could put the results of the stencil compiler in a
broken state.

debugging:

this bug is a race condition, and debugging + root causing was difficult
to consistently reproduce, as adding logging statements can alter the
built compiler and/or the results of the compilation step. through
adding the smallest amount of `console.trace` and `console.log`
statements to the build pipeline, starting with the build task, we were
able to confirm the following:
- only `.ts` files are responsible for causing the double compilation
  (which we'd later narrow down to `components.d.ts`)
- only the full build was being invoked (and not the watch task)
- `src/compiler/build/build.ts#build()` was invoked 2x, as evidenced by
  two 'Tranpile Started' messages in the output
- `afterProgramCreate()` was invoked 2x on the `WatchCompilerHost`
  (which confirms the `build` was running a second time in the previous
  bullet)
- the `fileUpdate` event was being invoked for a project's root
  `components.d.ts`, which caused the second `afterProgramCreate()` call
- the file update event was being created by this write to disk