docs(compiler): add jsdoc to the full build process
this commit adds jsdoc + context to existing code to improve clarity
between stencil, the ts compiler, and various ts apis. a few variables
were renamed as well to improve clarity.

this commit was separated from the actual fix to make the functional
changes more apparent, but the plan is to squash this commit into the
parent commit (that includes the fix)