chore(lint): turn on the eslint 'valid-typeof' rule
This turns on the eslint 'valid-typeof' rule which requires that any
comparisons using the `typeof` operator must be done in a valid way.

See here for details: https://eslint.org/docs/rules/valid-typeof

Adding because it is included in this list here:
https://github.com/typescript-eslint/typescript-eslint/blob/main/packages/eslint-plugin/src/configs/eslint-recommended.ts#L31

However, this constraint is already enforced by the Typescript compiler,
so we do not have any violations of it in the codebase! 🎉