chore(lint): add initial eslint jsdoc @param rules (#3390)
this commit adds the first pass at enforcing @param usage in jsdocs.
configuring how jsdoc's are enforced often involves setting up multiple
rules at once to get a desired affect. rather than toil with the
_perfect_ setup, this commit is designed to get an initial take on this
in front of the team to use and iterate on over time.

the choice of rules revolves around the general desire that each jsdoc
includes:
- the `@param` jsdoc keyword
- the name of the parameter
- a description of the parameter

it does not however:
- get enforced in every context (yet) - there are future discussions to
  decide whether arrow functions, anonymous functions, etc. are included
  or not
- get invoked if a context we _do_ enforce doesn't have a jsdoc (where
  we will continue to "clean up" the code as we go, adding jsdocs to
  functions we touch, just require `@param`s on them more formally)

the `.eslintrc.js` file is currently documented such that each
configuration option is explained/justified.