test(json-parsing): add unit tests  (#3398)
this commit adds testing around the `parsePackageJson` function. it
inlines a (former) internal function `parseJson`, which was originally
unnecessarily exported. Removing the `export` keyword from this
function was the impetus for this work, inlining the function came later
in the process

some types were tightened/changed:
- a concrete type, `ParsePackageJsonResult` was created to describe the
  return type of `parsePackageJson`. this type
  explicitly marks the `diagnostic` and `data` fields as the union of
  their original type and `null`.

this commit removes a check if the json that has been passed to
`parsePackageJson` is a string or not. looking at the single caller of it's
wrapper function, we verify that this argument is of type string
already.