feat(telemetry): add stencil config to telemetry object (#3401)
this adds an anonymized copy of the Stencil configuration to the
telemetry object, so that various configuration options can be tracked
and observed over time (while preserving anonymity). this is being done
in preparation for Stencil v3, so that we may get better to insights as
to which flags are being used (i.e. are they being used or not) in
order to determine what default flags we wish to set in v3

STENCIL-405: Add Stencil Configuration Contents to Telemetry