chore(tests): restore expectFiles functionality
this commit fixes a testing utility, `expectFiles`. prior to this
commit, the testing utility would call through to a file system
interface to verify that a file existed. however, the utility didn't do
anything with the results, as the return value of the file's existence
check:
1. didn't throw an `Error`
2. returned an object to designate whether the file existed or not
   (which was subesquently never queried)

this commit changes that behavior by `throw`ing in the event one or more
files cannot be found, and printing a message to the screen. `throw
Error()` was chosed over jest's `fail()` for 2 reasons:
1. it doesn't couple us to the jest testing library
2. (selfishly), it's easier to test

this commit _does_ add tests for the testing utility. at a certain
point, it does feel a little bit like overkill. however, this testing
utility is important in verifying that the output targets work properly.
the situtation of theh helper being broken for so long does give some
credence to having ~5 tests to verify it works properly