chore(lint): add lint for @returns
this commit enables multiple lint rules provided by the jsdoc plugin
regarding `@returns` in jsdocs.

the intent behind this commit is to ensure that jsdocs have a documented
`@returns` field and description where applicable, to catch missing ones
in ci rather than in code review.

there is one known limitation of these rules - when adding a jsdoc to a
function that also has a decorator applied:

```typescript
/**
 * A description
 * @returns the luckiest number in the world
 */
 @SomeDecorator()
 function getLuckyNumber(): number {
   return 7;
 }
```

the rule will error based on the return type of the decorator, rather
than the function.

there was only a single violation of this in all of stencil's code,
which was removed. i'm inclined to use this rule and use 'eslint-ignore'
directives if absolutely necessary, but would be open to arguments
against it