refactor(tests): update load-config tests
this commit updates the unit tests for loading a user configuration. it
does not seek to be comprehensive of the entire load pipeline. rather,
it seeks to improve the readability & assertions of existing tests.

this is being done as a part of a larger effort to attempt to make
configuration validation slightly stricter. by updating these tests
first, we can have greater confidence in changes that we shall make as a
part of the larger effort