chore(github): fix runtime setting in ci  (#3423)
this commit fixes a bug in our continuous integration pipeline where the
node runtime version was not being correctly set. for our analysis, e2e
and unit tests, we define a matrix of supported node versions to run on.
while the node version is used in the title of the step, it isn't used
anywhere else. furthermore, each step uses our 'get-core-dependencies'
action, which sets the version of the Node runtime to 16 (at the time of
this writing). it does not unset the version (overriding the default
version provided by github's hosted runners).

this commit updates the aforementioned workflows to run on the desired
version of node. the node version is set _after_ the dependencies are
retrieved to ensure the correct version is used.

note: this commit was designated as a 'chore' rather than 'fix' since
its not a user facing bug (and as such shouldn't appear in the changelog
for the release that contains it)