fix(styles): ensure styles are applied before paint
this commit fixes a bug where styles from a stencil component would not
be applied prior to render. the issue involved stylesheet construction &
application not blocking render, causing the component to be created
without styles applied.

this commit moves from the `replace` function to the `replaceSync`
function to properly await the call