refactor(style): refactor code in compiler/style/css-imports
This refactors and documents some code in the module responsible
for parsing css imports (src/compiler/style/css-imports.ts) and adds
some tests to increase coverage.

This commit also makes some edits to the general 'style' spec
(src/compiler/style/test/style.spec.ts) which has been skipped for a
while. The tests in this suite were 'end-to-end' unit tests of the style
handling capabilities of the compiler, which we'd like to have again,
but we can't reconstitute them until we get unit testing of the compiler
stood back up again.

So for now this ports a few tests out of `style.spec.ts` and into the
relevant specfiles (like `css-imports.spec.ts`) and deletes a few tests
from `style.spec.ts` which are no longer accurate (like testing for
escaping certain characters).