chore(test): restore doNotExpectFiles
this commit restores the `doNotExpectFiles` utility helper under test.
this commit is compliment to aee5287e1594086c0253030c03b62f9429d3ec7a,
which got the `expectFiles` testing utility under test.

this commit makes a few assumptions about how we will use this utility
function. namely, it only throws an exception if one or more files
provided by a caller are found. it does not check the permissions on the
file path (it appears that a previous implementation would take this
into account). should we decide to re-implement file access, we can do
so at a later date.

similar to aee5287e1594086c0253030c03b62f9429d3ec7a, this
re-implementation throws instead of using jest's `fail()` utility for
ease of testing the utility function (and doesn't couple the utility to
jest itself)