refactor(cli): refactor parse-flags module (#3427)
this refactors the parse-flags module, removing some code duplication
and changing how the `ConfigFlags` interface is declared.

Now the ConfigFlags interface is generated based on the content of a few
read-only arrays of strings. This means that config flags are listed in
one and only one placed and cannot drift / get out of sync.