chore(github): clean up github labels
this commit removes the 'reply received' label exemption from ionitron.
we stopped using this label ~3 months ago. in that time, the issues in
the backlog have all been re-triaged, and any issues with this label
today _also_ have an additional exempt label. this is the next step in
removing the label from the repo.

this also renames the 'reply received' label in the contributing guide
to 'awaiting reply' to align with the label we actually use.

the 'help wanted' label mention was removed, as we don't use it today