test(cli): add tests for run.ts (#3441)
this commit adds initial unit tests for `run.ts`. the goal of these
tests are to get the minimal amount of coverage under the `run` and
`runTask` functions so that we can (eventually) begin to change the
function signatures on the functions they invoke.

specifically, the stencil team is looking to make the configuration
object that is used throughout the compiler stricter by making fields
no longer optional on the type (internally). as a part of that effort, we're
reassessing what data fields functions need. we know that we will be
changing the signatures of the tasks for 'help' and 'telemetry'. by
adding tests around the invocation of these tasks, we can have greater
trust in the changes that we're making