refactor(config): add internal strict configuration type (#3436)
This PR proposes a methodology where we could start to make `Config` entities stricter, by swapping them out with an internal, stricter version. It introduces a new `ValidatedConfig` type that can gradually introduce strictness on the `Config` entity. For this PR, the only field that is made required is the `flags` field.

The `ValidatedConfig` type is used throughout the codebase following the validating of a user-supplied configuration. Public APIs cannot use `ValidatedConfig` without introducing a breaking change. As a result, an instance of `ValidatedConfig` is created and backfilled with the contents of the user provided config and any now necessary fields (defaulting them if the field is not provided).

In making `flags` a required field, the signatures of any function that accessed the field was updated to accept an `ValidatedConfig`. Not every instance of `Config` in a function signature was updated in this PR as a result.