chore(github): split browserstack into two jobs
this commit takes the pre-existing browserstack workflow (renamed to
`test-browserstack.yml`) and splits the job into two - one for building
stencil, and another to build the downstream applications + run
browserstack

the intent behind this commit is to make re-running browserstack a
faster operation for stencil engineers. rather than waiting
approximately 8 minutes to see if a job passes, we take the flakier of
the two (running browserstack) and move it into its own job. from the
github ui, we can "rerun failed jobs" for this workflow, resulting in a
~3.5 minute turnaround time. this is a part of a larger effort to
stabilize browserstack tests. while this commit doesn't perform any
stabilizations per se, it does help mitigate the time lost due to flaky
tests.

one of the downsides to this design is the level of repetion found in
our workflows. in the `test-browserstack.yml` workflow itself, there's
repetition in environment setup (installing dependencies, setting up
node, etc.). this unfortunately is the cost of splitting the workflow
into two jobs - some maintenance burden + repetition is necessary to be
able to re-trigger the browserstack tests in a shorter period of time.

similarly, there is repetition between the browserstack workflow and the
primary ci workflow; in both workflows, stencil core is built,
tarballed/archived, and pulled down by future steps. attempts to
coalesce these two were unsuccessful, as they either:
1. weakened our browserstack secret exposure
2. made the primary ci workflow more difficult to change in the future

as a result, the compromise of rebuilding stencil in an additional
github action runner was deemed acceptable by the author of this pr

STENCIL-433 Stabilize Browserstack CI Tests