chore(test): remove flaky test
this commit remvoes a flaky test that was discovered in refactoring the
existing test suites. a top level configuration entity was not being
properly reset between test runs, causing the test to fail.

looking at the test, it no longer tests a valid path in the codebase -
in running coverage before and after this change (with jest's cache
disabled), we see no decrease in code coverage