test(config): update tests for new validateConfig signature
this commit introduces a mock creation function, `mockLoadConfigInit` to
init the configuration file that used to bootstrap a task. this mock
creator is used throughout the test suite, in order to provide a
`LoadConfigInit` entity for `validateConfig`

this commit does surface the face that `validateConfig` is used
throughout the testing suite, and that the suite(s) should be refactored
to use more specific functions under test, rather than calling the
larger `validateConfig` function. this is being considered
'out-of-scope' & saved for a future day