refactor(config): add logger to `ValidatedConfig`
this commit adds the `Logger` instance to `ValidatedConfig`, requiring
the latter to hold an instance of the former.

the `validateConfig` function has been altered to accept an instance of
`LoadConfigInit`, the minimal configuration object that is used to
bootstrap a stencil task. this allows a central place for an unvalidated
configuration to be validated. Tather than semi-validating a config and
mutating it after the fact as a part of the load process (which is what
was previously done), all required fields are set in one place*

the `setupTestingConfig` function signature was altered to accept a
`ValidatedConfig` rather than a `Config`, as it's only caller has access
to a `ValidatedConfig`. this allows us to simplify some of the logic for
generating a validated config for testing purposes.

* for cases where the config does not pass through `validateConfig`,
such as `runTask`, we will attempt to pull a `Logger` off the
unvalidated config, or create one if one does not exist