chore(test): remove strictNullChecks from dev server validation (#3458)
this commit removes the strictNullCheck violations from tests for
validating the dev server. it only removes those associated with an
unvalidated configuration. we have no intention at the moment to make
`devServer` or its subfields required on `Config`. regardless of what
happens on `ValidatedConfig#devServer`, we needed to handle these
unvalidated `devServer` instances